backend/requirements/requirement.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-countries==7.6.1
django-extensions==4.1
django-filter==25.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
loguru==0.7.3
phonenumbers==9.0.7
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
shortuuid==1.0.13
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
